Data.Map.Base.generalMergeA: optimizing
Floated the case analysis on the MergeTactics out of the loops, so they are only performed once at the very beginning. Right now we still have a dispatch to a closure (generated by start2), but it's not clear how to eliminate that without lots of repetition of code. Still, should be faster than the version which performed case analysis on MergeTactic at every node.